+++ samples/webpack-app-hello-pug/src/app/app.js
@@ -1,0 +1,7 @@
+const html = require('./app.pug');
+const appRootSelector = '#app-root';
+
+export function app() {
+  const appRootContainer = document.getElementById(appRootSelector);
+  appRootContainer.innerHTML = html;
+}