+++ samples/webpack-app-hello-pug/src/main.js
@@ -1,0 +1,6 @@
+import { app } from './app/app';
+
+window.addEventListener('load', function (event) {
+  console.log('Start app!');
+  app();
+});