+++ test/cases/javascript-option-method-html/src/index.js
@@ -1,0 +1,3 @@
+const html = require(`Template/widget.pug?{"b":"yyy"}`);
+
+console.log(html);